Switched to new CLib NG repo

+++ src/iniReader.cpp
@@ -41,7 +41,7 @@ namespace INISettings {
 		if (range < 0.0) range = 0.0;
 		if (range > 150000.0) range = 150000.0;
 
-		ContainerManager::ContainerManager::GetSingleton()->SetMaxLookupRange(range);
+		ContainerManager::ContainerManager::GetSingleton()->SetMaxLookupRange((float) range);
 		return true;
 	}
 }
